chore: clippy/lint cleanup pass
- map_or -> is_none_or/is_some_and, if-let collapses, sort_by_key
- connect: dedicated thread + block_on, no std Mutex held across await
- dedicated.rs: build packet inside read guard, send after drop
- Default impls for Dispatcher/DispatchResult, PositionHistory::is_empty
- discovery: FNV_FIRST_ACTOR_LIST +0x00 is a no-op offset
- vtable: tail-expression null_mut (clippy::needless_return)
- hooks/mod: TODO markers for set_reputation/set_hardcore_stats

+++ crates/ashfall-server/src/world/cell.rs
@@ -12,7 +12,7 @@ impl CellGrid {
     /// Extract cell coordinates from a cell ID.
     /// Returns (world_id, x, y) for exterior cells, or (0, 0, 0) for interiors.
     pub fn decode_cell(cell: u32) -> (u32, i32, i32) {
-        let world_id = (cell >> 16) as u32;
+        let world_id = cell >> 16;
         let y = ((cell >> 8) & 0xFF) as i32;
         let x = (cell & 0xFF) as i32;
         (world_id, x, y)
